@thunderstore/cyberstorm: rename files for reusability
PackageSearch will have a second filter menu for "other" filters, so
rename files and dirs to faciliate adding it. It was not included in
the original changes, because it was missing in the old implementation
and I didn't realise I should account for it.

Refs TS-1860
