+++ packages/cyberstorm/src/components/PackageSearch/FilterMenus/CategoryMenu.tsx
@@ -3,7 +3,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import * as Checkbox from "@radix-ui/react-checkbox";
 import { Dispatch, SetStateAction } from "react";
 
-import styles from "./CategoryMenu.module.css";
+import styles from "./FilterMenu.module.css";
 import { CategorySelection, CATEGORY_STATES as STATES } from "../types";
 import { Icon } from "../../Icon/Icon";
